requirements.txt
faiss-cpu==1.10.0
langchain==0.3.23
langchain-community==0.3.21
langchain-openai==0.3.14
openai==1.75.0
python-dotenv==1.1.0
tiktoken==0.9.0